Applied patch sent in by Jaime frey
"it is a race condition in the job-manager between the main thread and a
query thread. The query function needs to signal graml_api_cond when it
sets graml_jm_can_exit back to false before it returns. Some of the
specific query types signal graml_api_cond in the middle of the query
function, but this is now redundant and can be removed. I haven't done
that, though."

-Stu