fixed for bug 813
rewrite the poll function to return job state DONE correctly.

Done is returned when LSF returns a state that maps to DONE or when the job is
not found by LSF.  Other cases like if LSF is down or bjobs returns an error,
the job manager will ignore the poll and keep the previously known state.