GRAM Client: Client parses RSL when making a job request: if username RSL attribute is present, modify the service name POSTed to the gatekeeper to indicate that the job should run as this user
Job Manager:
If username RSL attribute is present, validate that the job manager is running
as that user. If not, fail with an error (to catch old clients which don't
understand what to do with the RSL)
- add username to validation (rvf) file
- add validation call to state machine
Added missing prototype for one proxy-related function to eliminate warning.

Job Manager Test:
Added user test to submit jobs with good and bad username attributes.