Changes made in support of fail over.
myproxy-test-replicate was modified to use independent proxy when starting up the various servers.

myproxy.c had some bugs that were found using the new replicate test.  These were fixed.  As part of this fix, a new function retry_authentication_init was added.  There was also a lot of commented out code, this has been removed.

myproxy_authorization.c: The function auth_cert_check_client had a test that checked the owner tag and the proxy of the client.  This test was changed to test the client proxy and the subject name of the certificate in the repository.