adding a new program GridFTP Where There's FTP.
A security translator for clear text RFC959 to GSIFTP.  Currently it
only handles command translation on the control channel.  This should
suffice for 3rd party transfer.  We plane to add support for data
channel security and MODE S to MODE E translation.

Daemon mode is not yet complete, so currenlty 1 process must be run per
user.  Soon it can be a host service, so 1 process can service all users
on a host.