Work on GRAM-128: When restarting a crashed job manager, memory use can be high
This code changes the load_all functionality so that, after loading a job
request, depending on the job's restart_state, it either keeps track of the
job_contact_path for a periodic callback to process, or stows the job's LRM
job id in the job_id_hash so that the SEG (or fork SEG equivalent) will
reload the job when a state change occurs. Tested with fork (using
globus_gram_protocol_test, globus_gram_client_test, and
globus_gram_job_manager_test packages) and with an interactive scenario:
create 10 jobs, kill job manager, start job manager process. Committed for
additional testing on the OSG test machine