RIC-81: Combine threaded and nonthreaded builds
This first set of patches makes the detection of threading packages in core
non-exclusive (so that pthreads as well as none can be detected), and makes
it possible to have thread modules other than "none" loaded at runtime during
globus_i_thread_preactivate().

Added new API functions (though they need doc and test cases):
* globus_thread_set_model() to choose thread runtime model
* globus_thread_cancellable_func() to implement the pthread cancellation
  semantics without resorting to macros that must be paired in the interface.

Removed globus_thread_cleanup_push() and globus_thread_cleanup_pop() as they
can only be implemented as macros, which conflicts with the runtime choice
of thread models.

Removed the BUILD_LITE macro, so other parts of the toolkit will need to be
modified to use globus_i_am_only_thread() to determine whether it's reasonable
to create threads or not.

One test changes: globus_common_poll_test
The "random_stress" test had a test failure mode that was a race condition in
the test design. I changed that to be saner, though I'm not sure exactly
what the intention of the test is in that case.