Merge fixes for bug #7032 (From Mattias Ellert): The globus-openssl-module re-registers ssl object OIDs already present in the openssl library. This results in the same OID having more than one nid registration, which is a major source of proxy certificate verification failures.
This patch fixes this issue by checking if an OID is already registered beforeregistering it. It also changes places in the code where objects are referencesby their "short name" to do the references by OID, since the OIDs are the samein globus and openssl, but the short names are not. The globus-gsi-proxy-core
package also needs some changes to use the openssl encoding functions when the
openssl registrations are used.