Minor fixes from Mattias Ellert: - grid-packaging-tools-pkgconfig.patch
1) All the .pc files say "Name: gptname" due to a missing $ sign
2) Change "Requires" to "Requires.private", see
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=612524
3) $extlibs will always contain @GPT_EXTERNAL_LIBS@, so the "if
($extlibs ne '')" doesn't make sense.

- grid-packaging-tools-escape.patch

Make gpt work if the build directory path contains characters that have
special meaning in a perl s/// expression. Doesn't happen so often on
Fedora, but in debian rebuilds of a package without source package
changes (to pick up a soname bump in one of its depenencies for example)
usually have +b1 added to the version, and the path to the temporary
build directory contains the package version in the debian build system.
You also often get many + signs in temporary paths on MacOS X for some
reason.

- globus-core-env.patch

1) The @STATIC_LDFLAGS@ substitution is never defined, so this string
with the @ signs remain in the installed files.
2) The 'test "$exec_prefix" = NONE' is not needed here.

- globus-core-spec-creator.patch

Several different problems with the globus-spec-creator script.