First draft of code to allow gatekeeper/jobmanager communications to work in absence of a working gss_export_sec_context/gss_import_sec_context pair
The gatekeeper has a new launch method FORK_AND_PROXY which causes it to
fork the job manager process and wait for the reply and wrap it itself.
It also changes the amount and type of info it communicates with the job
manager via environment variables.

Most of the new variables are based off of apache's mod_ssl and CGI
implementations, so in theory the job manager could be pushed out via
other methods (though the fork and handoff between processes might confuse
a true CGI server).

Most of the main behavior is spot-tested on MacOS by explicitly setting the
launch method.