Do not destroy global logging handle.
The destruction of the global logging handle in the gridftp server
is not thread-safe.  When in threaded mode the logging handle can be
used it is closed.  Since there is no concurrency management of the
global pointer (either atomics or mutexes), we cannot set it to null.

Felt that a layer of global mutexes isn't the right approach and
didn't see too many other examples of atomics in Globus - hence,
just leak the resource.