Allow sysconf(_SC_GETPW_R_SIZE_MAX) to return -1
On FreeBSD, sysconf(_SC_GETPW_R_SIZE_MAX) returns -1.  This is a valid
return value (it means we need to guess what buffer size to use), but
right now we just assert if -1 is returned.

With this change, if the sysconf() returns a value, we use that as
the starting size for a malloc().  If sysconf() returns -1, we use
1024.  If either size ends up being too small, we free() and then
try again with a larger size (increasing in chunks of 1024).